# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Export Client Peer API
  version: 2.0.0
  description: >-
    API demo untuk tiga backend mandiri (web, mobile, desktop). Inquiry direplikasi
    secara append-only antar-peer; tidak ada master, update, atau delete.
servers:
  - url: http://localhost:4000
    description: Backend web lokal (default)
paths:
  /health:
    get:
      summary: Memeriksa kesehatan backend dan database lokal
      responses:
        '200':
          description: API siap
          content:
            application/json:
              schema:
                type: object
                required: [status, nodeId, syncEnabled]
                properties:
                  status: { type: string, const: ok }
                  nodeId: { type: string }
                  syncEnabled: { type: boolean }
  /api/v1/products:
    get:
      summary: Mendapatkan katalog produk aktif dari database lokal
      responses:
        '200':
          description: Katalog demo lokal
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, name, category, origin, unit]
                      properties:
                        id: { type: string }
                        name: { type: string }
                        category: { type: string }
                        origin: { type: string }
                        unit: { type: string }
  /api/v1/inquiries:
    post:
      summary: Membuat inquiry pada database backend lokal
      description: >-
        Inquiry baru mendapat ID UUID dan kode pelacakan stabil. Bila sinkronisasi
        diaktifkan, record append-only disebarkan ke peer yang dikonfigurasi.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerName, customerEmail, destinationCountry, productId, quantity]
              properties:
                customerName: { type: string, minLength: 2, maxLength: 120 }
                customerEmail: { type: string, format: email, maxLength: 254 }
                destinationCountry: { type: string, minLength: 2, maxLength: 80 }
                productId: { type: string }
                quantity: { type: number, exclusiveMinimum: 0, maximum: 1000000 }
      responses:
        '201':
          description: Inquiry dibuat; kode berlaku pada semua peer setelah direplikasi
          content:
            application/json:
              schema:
                type: object
                required: [trackingCode, status, createdAt]
                properties:
                  trackingCode: { type: string, pattern: '^[A-F0-9]{24}$' }
                  status: { type: string, enum: [received] }
                  createdAt: { type: string, format: date-time }
        '400': { description: JSON tidak valid atau validasi input gagal }
        '404': { description: Produk tidak ditemukan }
        '429': { description: Batas RFQ per 15 menit tercapai }
        '413': { description: JSON request body melebihi 16 KiB }
  /api/v1/inquiries/{trackingCode}:
    get:
      summary: Melacak status inquiry di database peer lokal
      description: >-
        Memerlukan kode bearer 24 karakter. Respons tidak memuat nama atau email.
        Replication eventual; tunggu sinkronisasi bila inquiry dibuat di peer lain.
      parameters:
        - in: path
          name: trackingCode
          required: true
          schema: { type: string, pattern: '^[A-F0-9]{24}$' }
      responses:
        '200':
          description: Status inquiry lokal
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status, createdAt, productName]
                    properties:
                      status: { type: string, enum: [received] }
                      createdAt: { type: string, format: date-time }
                      productName: { type: string }
        '400': { description: Format kode tidak valid }
        '404': { description: Inquiry belum tiba di peer atau tidak ditemukan }
        '429': { description: Batas lookup per 15 menit tercapai }
  /api/v1/sync/inquiries:
    post:
      summary: Menerima atau mereplikasi record inquiry peer
      description: >-
        Endpoint layanan-ke-layanan. Nonaktif bila SYNC_SHARED_SECRET kosong.
        Record bersifat append-only: ID/kode duplikat identik no-op; isi berbeda
        menghasilkan 409 dan disimpan sebagai konflik tanpa menimpa record lokal.
        Gunakan TLS untuk sinkronisasi di luar loopback; jangan kirim data nyata.
      security:
        - PeerSyncBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InquiryRecord' }
      responses:
        '201': { description: Record baru diterima }
        '200': { description: Record identik sudah tersimpan }
        '400': { description: JSON tidak valid atau record tidak valid }
        '401': { description: Bearer secret tidak sah }
        '404': { description: Produk demo tidak tersedia pada peer }
        '409':
          description: Konflik; record lokal tidak berubah
          content:
            application/json:
              schema:
                type: object
                required: [error, result, existingHash, incomingHash]
                properties:
                  error: { type: string, const: sync_conflict }
                  result: { type: string, const: conflict }
                  existingHash: { type: string }
                  incomingHash: { type: string }
        '503': { description: Sinkronisasi belum diaktifkan di server penerima }
        '413': { description: JSON request body melebihi 16 KiB }
components:
  securitySchemes:
    PeerSyncBearer:
      type: http
      scheme: bearer
      description: Secret antarlayanan minimal 32 karakter; jangan dibundel di klien.
  schemas:
    InquiryRecord:
      type: object
      additionalProperties: false
      required: [id, trackingCode, customerName, customerEmail, destinationCountry, productId, quantity, unit, status, createdAt, originNodeId]
      properties:
        id: { type: string, format: uuid }
        trackingCode: { type: string, pattern: '^[A-F0-9]{24}$' }
        customerName: { type: string, minLength: 2, maxLength: 120 }
        customerEmail: { type: string, format: email, maxLength: 254 }
        destinationCountry: { type: string, minLength: 2, maxLength: 80 }
        productId: { type: string, minLength: 1, maxLength: 80 }
        quantity: { type: number, exclusiveMinimum: 0, maximum: 1000000 }
        unit: { type: string, minLength: 1, maxLength: 20 }
        status: { type: string, enum: [received] }
        createdAt: { type: string, format: date-time }
        originNodeId: { type: string, pattern: '^[A-Za-z0-9._-]{1,80}$' }